add password health service
